Improve portability: drop `readlink` for pubsubc and test scripts
BSD version of `readlink`, which is the default on macOS, doesn't have
'-f, --canonicalize' option. And without this option we might just as
well use a dummy approach, of dropping the last element from the
script's path, to detect its base directory. Though, it doesn't work
when script is executed from $PATH, it fully suits this simple PoC
needs.

pubsubd script can continue using `readlink -f` because it runs in
docker debian-based container, where GNU readlink is the default.