Fix trap handlers in the test script
In Bash, if multiple `trap` handlers are set for the same signal
they're not accumulated, but the more recent one simply overwrites the
existing handler.

In order to be able to stack several trap handlers, we read the current
handler value and inject it into the new handler, as a separate command.